Fix parent broker connection heartbeat interval
The heartbeat timer should be reset only when a message (or
a frame) is sent to the parent broker, but not when a frame
or another event is received.
In a case where there are only frames coming from the parent
broker, but no frames are being send, the timer would get
reset on each receive and no heartbeat would be sent, which
could result in disconnect from the parent broker.