Add `AccessLevel` meta-attribute for RpcMessage
A new attribute to specify access level in requests by a numerical
representation.

The older `Access` meta-attribute, which uses a string representation
is kept for compatibility with older brokers/clients as a fallback
if `AccessLevel` is not present.

The `Access` to `AccessLevel` conversion is performed as follows:

If the `Access` string contains a comma-separated list of grants
the first known grant that a conversion from string succeeds for
is returned, the rest is ignored.

(#14)