Provide backward compatibility for the dot-local hack
Broker sets both 'Access' and 'AccessLevel' RpcMessage meta
attributes or pass them as is from a parent broker in order
to keep backward compatibility with older brokers that read
'dot-local' in the 'Access' string to expose their '.local'
dir.

The AccessLevel is treated as plain i32, so broker can evaluate
access also for values that are not defined in the enum.

(#14)