Implement `From` on generic collections for `RpcValue`
Add blanket implementations of traits:

 - From<Vec<T>>
 - From<BTreeMap<String, T>>
 - From<BTreeMap<i32, T>>

for RpcValue where T: Into<RpcValue>

This allows to perform uniform conversion to RpcValue for
structs whose fields are either of a type convertible to
RpcValue or a collection of such type.

There is no specialization for T equals to `RpcValue` due
to limitations of Rust, so the conversion is unnecessarily
costly as it iterates through the collection instead of just
moving the inner data.
The compiler in release mode is able to generate optimized code
of `From<Vec<RpcValue>>`, but unfortunately not for `BTreeMap`.