TryFromRpcValue: Provide an attribute for custom tag
A user can attach an attribute to an enum and define which
key should be used as the enum variant discriminant.

Tags currently work only for enums with named fields and
only internal tagging is implemented. A tag value is
serialized to the map as enum variant converted to camelCase.

If no tag is explicitly defined, the value defaults to "type".

Example:

 #[derive(TryFromRpcValue)]
 #[rpcvalue(tag = "pet")]
 enum Animal {
     Cat { name: String, age: i32, cuteness_level: i32 },
     Dog { name: String, age: i32, barking_loudness: i32 },
 }

A Cat variant will be converted to RpcValue as:

 { "pet": "cat", "name": "meoow", age: 365, cuteness_level: 100 }